hack: Rewrite injecting GSX irq logic
Handle GSX irq separately from other guest irqs. Now Xen subscribes to
GSX irq and performs all GSX related actions in an extra interrupt handler.
Xen reads both irq counters (host and guest) and injects GSX irqs
to required domains.
