+++ xen/arch/arm/irq.c
@@ -26,7 +26,6 @@
 
 #include <asm/gic.h>
 #include <asm/vgic.h>
-#include <asm/io.h>
 
 static unsigned int local_irqs_type[NR_LOCAL_IRQS];
 static DEFINE_SPINLOCK(local_irqs_type_lock);
@@ -183,17 +182,6 @@ int request_irq(unsigned int irq, unsigned int irqflags,
     return retval;
 }
 
-#define GSX_IRQ_NUM                151
-#define GSX_REG_BASE               0xfd000000
-#define GSX_REG_SIZE               0x00002000
-#define GSX_IRQ_CNTS_REQ           0x00001238
-#define GSX_IRQ_CNTS_SHIFT         12
-#define GSX_GUEST_IRQ_CNT_SHIFT    14
-
-static void __iomem *gsx_reg_base = NULL;
-
-static u16 guest_irq_cnt;
-
 /* Dispatch an interrupt */
 void do_IRQ(struct cpu_user_regs *regs, unsigned int irq, int is_fiq)
 {
@@ -235,39 +223,7 @@ void do_IRQ(struct cpu_user_regs *regs, unsigned int irq, int is_fiq)
          * The irq cannot be a PPI, we only support delivery of SPIs to
          * guests.
 	 */
-        if ( irq != GSX_IRQ_NUM )
-            vgic_vcpu_inject_spi(info->d, info->virq);
-        else
-        {
-            struct domain *d;
-            u32 irq_cnts;
-
-            if ( !gsx_reg_base )
-            {
-                gsx_reg_base = ioremap_nocache(GSX_REG_BASE, GSX_REG_SIZE);
-                if ( !gsx_reg_base )
-                    printk("failed to map GSX MMIO range\n");
-            }
-
-            irq_cnts = readq_relaxed(gsx_reg_base + GSX_IRQ_CNTS_REQ) >> GSX_IRQ_CNTS_SHIFT;
-
-            vgic_vcpu_inject_irq(info->d->vcpu[0], GSX_IRQ_NUM);
-
-            if ( guest_irq_cnt != (irq_cnts >> GSX_GUEST_IRQ_CNT_SHIFT) )
-            {
-                for_each_domain ( d )
-                {
-                    if ( d->domain_id < 2 )
-                        continue;
-
-                    if ( likely(d->vcpu != NULL) && likely(d->vcpu[0] != NULL) )
-                        vgic_vcpu_inject_irq(d->vcpu[0], GSX_IRQ_NUM);
-                    break;
-                }
-                guest_irq_cnt = irq_cnts >> GSX_GUEST_IRQ_CNT_SHIFT;
-            }
-        }
-
+        vgic_vcpu_inject_spi(info->d, info->virq);
         goto out_no_end;
     }
 
@@ -515,8 +471,6 @@ int route_irq_to_guest(struct domain *d, unsigned int virq,
      *  - If it's the same domain -> Xen doesn't need to update the IRQ desc.
      *  For safety check if we are not trying to assign the IRQ to a
      *  different vIRQ.
-     *  - Otherwise -> For now, don't allow the IRQ to be shared between
-     *  Xen and domains.
      */
     if ( desc->action != NULL )
     {
@@ -526,12 +480,9 @@ int route_irq_to_guest(struct domain *d, unsigned int virq,
 
             if ( d != ad )
             {
-                if ( irq != GSX_IRQ_NUM )
-                {
-                    printk(XENLOG_G_ERR "IRQ %u is already used by domain %u\n",
-                           irq, ad->domain_id);
-                    retval = -EBUSY;
-                }
+                printk(XENLOG_G_ERR "IRQ %u is already used by domain %u\n",
+                       irq, ad->domain_id);
+                retval = -EBUSY;
             }
             else if ( irq_get_guest_info(desc)->virq != virq )
             {
@@ -540,11 +491,6 @@ int route_irq_to_guest(struct domain *d, unsigned int virq,
                        d->domain_id, irq, irq_get_guest_info(desc)->virq);
                 retval = -EBUSY;
             }
-        }
-        else
-        {
-            printk(XENLOG_G_ERR "IRQ %u is already used by Xen\n", irq);
-            retval = -EBUSY;
         }
         goto out;
     }
