+++ xen/arch/arm/setup.c
@@ -804,6 +804,7 @@ void __init start_xen(unsigned long boot_phys_offset,
     xsm_dt_init();
 
     init_maintenance_interrupt();
+    init_gsx_interrupt();
     init_timer_interrupt();
 
     timer_init();
